containers/backend/requirements.txt
alembic==1.13.1
annotated-types==0.6.0
anyio==4.3.0
click==8.1.7
fastapi==0.110.0
greenlet==3.0.3
gunicorn==21.2.0
h11==0.14.0
idna==3.6
Mako==1.3.2
MarkupSafe==2.1.5
packaging==24.0
psycopg2-binary==2.9.9
pydantic==2.6.4
pydantic_core==2.16.3
python-dotenv==1.0.1
pytz==2024.1
sniffio==1.3.1
SQLAlchemy==2.0.29
starlette==0.36.3
typing_extensions==4.10.0
uvicorn==0.29.0
